[PLAT-569] rename entry point
