+++ src/device-uptime/Dockerfile
@@ -10,11 +10,11 @@ WORKDIR /app/
 RUN pip install -r requirements.txt
 # Set "python" as the entry point
 # Set environment variables
-ENV FLASK_ENV=production
+ENV ENV=production
 ENV GOOGLE_APPLICATION_CREDENTIALS=airqo-250220-642e046f4223.json
 
 # Expose the application's port, default is 5000 for python flask
 EXPOSE 5000
 
 # Run the application
-CMD ["python", "app.py", "--tenant=airqo"]
+CMD ["python", "main.py", "--tenant=airqo"]
