+++ src/device-uptime/utils/device_hourly_records.py
